requirements.txt
numpy<2
einops
tqdm
ipython
dm_control==1.0.11
mujoco==2.3.3
gym
gymnasium[all]==0.29.1
gymnasium[accept-rom-license]
tensorboard
omegaconf
opencv-python
hydra-core
moviepy
timm==0.6.12
termcolor
h5py
imageio==2.27 # bug in 2.28 make the gif still
xformers==0.0.22
cython<3 # some bug be triggered for mujoco-py build for cython3
tensordict==0.2
pot
torchinfo